msm: mdss: fix performance during continuous backlight updates
Commit f914743bccfe ("msm: mdss: sync mdp kickoff and non-burst dsi
commands") introduced a regression in performance when multiple
backlight commands are getting sent along with MDP data.
The fix intended to fix race condition when both MDP and DSI command
transfer wait for MDP transfer to finish. Since DSI commands can go in
parallel (with burst mode), it's better to treat as such. Add a mutex to
explicitly control exclusive access to MDP, that is, appropriately wait
on MDP to be idle if needed by holding this mutex. This guarantees that
DSI commands that need to be sent while MDP is idle will remain idle
until transfer of commands is complete.

Fixes: f914743bccfe
Bug: 63169931
Signed-off-by: Adrian Salido <salidoa@google.com>
Change-Id: Iabbb671fe7540c7e19a5f80e57c7447ae745c15d
Signed-off-by: Dede Dindin Qudsy <xtrymind@gmail.com>